+++ src/components/choropleth2/tooltips/municipal/createMunicipalHospitalAdmissionsTooltip.tsx
@@ -1,8 +1,8 @@
 import { NextRouter } from 'next/router';
 import { ReactNode } from 'react';
-import { MunicipalityProperties } from '~/components/choropleth/shared';
-import { createSelectMunicipalHandler } from '~/components/choropleth/selectHandlers/createSelectMunicipalHandler';
-import { TooltipContent } from '~/components/choropleth/tooltips/tooltipContent';
+import { MunicipalityProperties } from '~/components/choropleth2/shared';
+import { createSelectMunicipalHandler } from '~/components/choropleth2/selectHandlers/createSelectMunicipalHandler';
+import { TooltipContent } from '~/components/choropleth2/tooltips/tooltipContent';
 
 export const createMunicipalHospitalAdmissionsTooltip = (
   router: NextRouter
