+++ src/components/choropleth2/tooltips/municipal/createPositiveTestedPeopleMunicipalTooltip.tsx
@@ -2,7 +2,7 @@ import { NextRouter } from 'next/router';
 import { ReactNode } from 'react';
 import { MunicipalityProperties } from '../../shared';
 import { createSelectMunicipalHandler } from '../../selectHandlers/createSelectMunicipalHandler';
-import { TooltipContent } from '~/components/choropleth/tooltips/tooltipContent';
+import { TooltipContent } from '~/components/choropleth2/tooltips/tooltipContent';
 import { replaceVariablesInText } from '~/utils/replaceVariablesInText';
 import siteText from '~/locale/index';
 const text = siteText.common.tooltip;
