+++ src/components/choropleth2/tooltips/region/escalationTooltip.tsx
@@ -4,8 +4,8 @@ import EscalationLevel1 from '~/assets/niveau-1.svg';
 import EscalationLevel2 from '~/assets/niveau-2.svg';
 import EscalationLevel3 from '~/assets/niveau-3.svg';
 import EscalationLevel4 from '~/assets/niveau-4.svg';
-import { regionThresholds } from '~/components/choropleth/regionThresholds';
-import { TooltipContent } from '~/components/choropleth/tooltips/tooltipContent';
+import { regionThresholds } from '~/components/choropleth2/region-thresholds';
+import { TooltipContent } from '~/components/choropleth2/tooltips/tooltipContent';
 import text from '~/locale/index';
 import { formatDateFromSeconds } from '~/utils/formatDate';
 import { replaceVariablesInText } from '~/utils/replaceVariablesInText';
